% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ppa.R
\name{generate_ppa}
\alias{generate_ppa}
\alias{generate_ppa.data.frame}
\title{Generic Function to Calculate Presences and Pseudo-absences from an
Enviromental Space ES and presences P}
\usage{
generate_ppa(ES, P, ES_idx, ppa_function, remove_ES_idx, ...)

\method{generate_ppa}{data.frame}(
  ES,
  P = NULL,
  ES_idx = NULL,
  ppa_function = NULL,
  remove_ES_idx = TRUE,
  ...
)
}
\arguments{
\item{ES}{Um dataframe representando o espaço ambiental contendo as variáveis a
serem usadas na geração das pseudo-ausências. Esse dataframe pode conter colunas
representando as presenças das espécies. Cada espécie em uma coluna e cada
presença em uma linha denotada por valores maiores do que zero. Pode conter,
também,  uma coluna indicando o \code{ES_idx}, que é um índice para o espaço
geográfico representado, que pode referenciar uma grid
de polígonos ou um raster. Se houver \code{ES_idx} em \strong{ES}, existe a opção de
mante-lo ou não nas presenças e pseudo-ausências geradas.}

\item{P}{Caso \strong{ES} contenha colunas representando as presenças, \strong{P} pode
ser um vetor com os nomes das colunas que representam as presenças das
espécies.
Caso \strong{ES} contenha apenas o espaço ambiental, \strong{P} é um dataframe que
representa as presenças das espécies. Cada espécie em uma coluna e cada
presença em uma linha denotada por valores maiores do que zero. \strong{P} pode
conter uma coluna \code{ES_idx}, indicando qual linha em \strong{P} será ligada a qual
linha de \strong{ES} por meio de uma juncão. Mas, se não houver e a quantidade de
linhas em \strong{P} for igual a quantidade de linhas em \strong{ES}, ao invés de um join,
um \code{bind_cols} é usado para indicar qual linha em \strong{P} é ligada a qual linha
em \strong{ES}. Se \code{ES_idx} não estiver em \strong{P} e quantidade de linhas em \strong{P} for
diferente de \strong{ES}, um erro é gerado.}

\item{ES_idx}{Indica o nome da coluna em \strong{ES}, e eventualmente em \strong{P}, é
usada para indicar o espaço geográfico representado em \strong{ES} e, também, para
ligar \strong{ES} a \strong{P}. Pode ser omitido e, nesse caso, o espaço geográfico é
desconsiderado.}

\item{ppa_function}{Função que é responsável por fazer a geração das
presenças e pseudo-ausências. Essa função é passada para
\code{.generate_single_ppa} e deve seguir os mesmos parâmetros que
\code{.generate_single_ppa.}}

\item{remove_ES_idx}{Indica se a coluna que representa o índice do espaço
geográfico deve ou não ser mantido nas presenças e pseudo-ausências geradas.}

\item{...}{Demais parâmetros passadas para a função \code{ppa_function}.}
}
\value{
O valor retornado é uma lista nomeada com as espécies e, em cada
posição da lista, um dataframe contendo as presenças e pseudo-ausências,
juntamente com as variáveis ambientais e, opcionalmente, \code{ES_idx}.
}
\description{
Essa função é usada por todos os métodos de geração de pseudo-ausências. Seu
objetivo é validar os dados de entrada para tal geração, execuar a geração
para cada uma das espécies e, validar a saída grenéria para cada método de
geração.
}
\examples{

\dontrun{
a
}
}
