% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppa_cluster.R
\name{ppa_cluster}
\alias{ppa_cluster}
\alias{ppa_cluster.data.frame}
\title{Calcula as Presenças e Pseudo-Ausências usando Clusterização sobre o
Espaço Ambiental ES e Presenças P}
\usage{
ppa_cluster(ES, P, ES_idx, remove_ES_idx, num_clusters, perc_pres)

\method{ppa_cluster}{data.frame}(
  ES,
  P = NULL,
  ES_idx = NULL,
  remove_ES_idx = TRUE,
  num_clusters = 0,
  perc_pres = 10
)
}
\arguments{
\item{ES}{A set representing the environmental space.}

\item{P}{Presences}

\item{ES_idx}{Index}

\item{remove_ES_idx}{Removing idx}

\item{num_clusters}{Clusters}

\item{perc_pres}{Removing}
}
\value{
PPA
}
\description{
\code{ppa_cluster} calculates a set of pseudo-absences points (PA) \strong{randomly}
drawn from the enviromental space denoted by ES.
}
\details{
Some details
}
\examples{
P <- dplyr::tibble(
  cell_id = c(
    75000, 81017, 81374, 83506, 83508,
    84360, 84783, 84785, 86047, 86468
  ),
  bio_1 = c(
    25.24835, 26.47237, 25.09730, 25.28250, 25.25749,
    25.43329, 25.33722, 25.36212, 25.50991, 25.84756
  ),
  bio_2 = c(
    12.365706, 10.976032, 11.488524, 11.492427, 11.356132,
    11.477320, 11.125202, 11.271169, 10.878525, 10.985706
  ),
  bio_3 = c(
    68.53027, 66.84972, 70.87223, 70.63369, 70.99750,
    70.81410, 70.95460, 71.04424, 71.87370, 72.48579
  ),
  bio_4 = c(
    48.37795, 54.44825, 42.10769, 42.21368, 40.86964,
    41.39428, 41.72614, 41.17366, 44.82830, 46.00434
  ),
  sp1 = c(0,0,0,1,1,1,0,1,0,0)
)

PPA <- P |> ppa_cluster(c("sp1"))

}
