% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmpap-package.R
\docType{package}
\name{sdmppa-package}
\alias{sdmppa}
\alias{sdmppa-package}
\title{sdmppa: Presences and Pseudo-absences generation for Species Distribution Modeling}
\description{
The package generates pseudo-absence points (PAP) to be used in species distribution modeling (SDM) and environmental niche modeling (ENM). The process involves taking an environmental space represented by a set of points P and a set of presence points (PP) as input and then generating points that mimic environmental conditions of a true absence. The resulting PPA is a subset of P and does not contain any points in common with PP, meaning that PP and PPA are disjoint sets. The package provides several different methods of PPA generation.
}
\author{
\strong{Maintainer}: Reginaldo Ré \email{reginaldo@utfpr.edu.br} (\href{https://orcid.org/0000-0001-6452-3466}{ORCID})

Authors:
\itemize{
  \item Dayani Bailly \email{dayanibailly@nupelia.uem.br} (\href{https://orcid.org/0000-0002-6954-9902}{ORCID})
  \item Luiz Fernando Esser \email{luizesser@gmail.com} (\href{https://orcid.org/0000-0003-2982-7223}{ORCID})
  \item Marcos Robalinho Lima \email{robalinho@uel.br} (\href{https://orcid.org/0000-0002-5901-0911}{ORCID})
}

}
\keyword{internal}
