% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ppa.R
\name{generate_ppa}
\alias{generate_ppa}
\alias{generate_ppa.data.frame}
\title{Generic Function to Calculate Presences and Pseudo-absences from an
Enviromental Space ES and presences P}
\usage{
generate_ppa(ES, P, ES_idx, ppa_function, ...)

\method{generate_ppa}{data.frame}(ES, P = NULL, ES_idx = NULL, ppa_function = NULL, ...)
}
\arguments{
\item{ppa_function}{}
}
\value{
\code{NULL}
}
\description{
Generic Function to Calculate Presences and Pseudo-absences from an
Enviromental Space ES and presences P
}
