% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppa_random.R
\name{ppa_random}
\alias{ppa_random}
\alias{ppa_random.data.frame}
\title{Calculate Randomly Presences and Pseudo-absences from an Enviromental Space ES and presences P}
\usage{
ppa_random(ES, P, ES_idx)

\method{ppa_random}{data.frame}(ES, P = NULL, ES_idx = NULL)
}
\arguments{
\item{ES}{}

\item{ES_idx}{}
}
\value{
\code{NULL}
}
\description{
Calculate Randomly Presences and Pseudo-absences from an Enviromental Space ES and presences P
}
